requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
numpy==1.26.4
opencv-python==4.9.0.80
pillow==12.0.0
requests==2.32.5
tflite-runtime==2.14.0
urllib3==2.5.0
